Closes #12
Task Outcome: The task has been completed with all requested functionality. The instructor should expect to find an index of job postings, show pages for each posting, the functionality to create a new job posting, and, while signed in as a recruiter, the ability to delete a job posting they made.